+++ src/Authorization/AuthorizationService.php
@@ -7,6 +7,7 @@ namespace Dbp\Relay\FormalizeBundle\Authorization;
 use Dbp\Relay\AuthorizationBundle\API\ResourceActionGrantService;
 use Dbp\Relay\CoreBundle\Authorization\AbstractAuthorizationService;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
+use Dbp\Relay\CoreBundle\Rest\Query\Pagination\Pagination;
 use Dbp\Relay\FormalizeBundle\Entity\Form;
 use Dbp\Relay\FormalizeBundle\Entity\Submission;
 
@@ -139,13 +140,24 @@ class AuthorizationService extends AbstractAuthorizationService
     }
 
     /**
+     * If $firstResultIndex is 0 and $maxNumResults null, all results are returned.
+     *
      * @return string[][]
      */
-    public function getGrantedFormActionsPage(array $whereActionsContainOneOf,
-        int $firstResultIndex = 0, int $maxNumResults = self::MAX_NUM_RESULTS_MAX): array
+    public function getGrantedFormActions(array $whereActionsContainOneOf,
+        int $firstResultIndex = 0, ?int $maxNumResults = null): array
     {
-        return $this->resourceActionGrantService->getGrantedItemActionsPageForCurrentUser(self::FORM_RESOURCE_CLASS,
-            $whereActionsContainOneOf, $firstResultIndex, $maxNumResults);
+        if ($firstResultIndex === 0 && $maxNumResults === null) {
+            return Pagination::getAllResults(
+                function (int $currentPageStartIndex, int $maxNumItemsPerPage) use ($whereActionsContainOneOf) {
+                    return $this->resourceActionGrantService->getGrantedItemActionsPageForCurrentUser(
+                        self::FORM_RESOURCE_CLASS, $whereActionsContainOneOf,
+                        $currentPageStartIndex, $maxNumItemsPerPage);
+                });
+        } else {
+            return $this->resourceActionGrantService->getGrantedItemActionsPageForCurrentUser(self::FORM_RESOURCE_CLASS,
+                $whereActionsContainOneOf, $firstResultIndex, $maxNumResults ?? self::MAX_NUM_RESULTS_MAX);
+        }
     }
 
     public function isCurrentUserAuthorizedToCreateForms(): bool